Support dispatching non-reentrant events from a node as new events into the top level of the graph
